qemu config
  Конфигурация для запуска qemu

+++ Sources/proj_cmake/main.c
@@ -1,8 +1,9 @@
 #include "main.h"
 
+
 int main(void){
-int a = 123;
-int i =0;
+uint16_t a = 123;
+uint16_t i =0;
     while(i < 5){
         printf("%d",a);
         i++;
